move functions to ComponentSpecs, remove mixin from createI13nNode
