+++ src/utils/setupI13n.js
@@ -4,10 +4,21 @@
  */
 'use strict';
 
+var ComponentSpecs = require('../libs/ComponentSpecs');
 var React = require('react');
 var ReactI13n = require('../libs/ReactI13n');
-var I13nUtils = require('../mixins/I13nUtils');
+var hoistNonReactStatics = require('hoist-non-react-statics');
 var IS_CLIENT = typeof window !== 'undefined';
+
+function mixin(target, source) {
+    target = target.prototype;
+    source = source;
+    Object.getOwnPropertyNames(source).forEach(function (name) {
+        if (name !== "constructor") {
+            Object.defineProperty(target, name, Object.getOwnPropertyDescriptor(source, name));
+        }
+    });
+}
 
 /**
  * Create an app level component with i13n setup
@@ -24,22 +35,13 @@ var IS_CLIENT = typeof window !== 'undefined';
 module.exports = function setupI13n (Component, options, plugins) {
     options = options || {};
     plugins = plugins || [];
-    var RootI13nComponent;
-    var componentName = Component.displayName || Component.name;
 
-    RootI13nComponent = React.createClass({
-
-        mixins: [I13nUtils],
-
-        displayName: options.displayName || ('RootI13n' + componentName),
-
-        autobind: false,
-
-        /**
-         * componentWillMount
-         * @method componentWillMount
-         */
-        componentWillMount: function () {
+    class RootI13nComponent extends React.Component {
+        constructor(props) {
+            super(props);
+        }
+        
+        componentWillMount () {
             var reactI13n = new ReactI13n(options);
             this._reactI13nInstance = reactI13n;
             // we might have case to access reactI13n instance to execute event outside react components
@@ -51,9 +53,9 @@ module.exports = function setupI13n (Component, options, plugins) {
                 reactI13n.plug(plugin);
             });
             reactI13n.createRootI13nNode();
-        },
+        }
 
-        render: function () {
+        render () {
             var props = Object.assign({}, this.props);
             if (!options.skipUtilFunctionsByProps) {
                 props.i13n = {
@@ -68,7 +70,24 @@ module.exports = function setupI13n (Component, options, plugins) {
                 props.children
             );
         }
+    }
+    
+    mixin(RootI13nComponent, ComponentSpecs.pickSpecs([
+        'getChildContext',
+        'executeI13nEvent',
+        'getI13nNode',
+        '_getReactI13n',
+        '_getParentI13nNode',
+    ]));
+    
+    RootI13nComponent = Object.assign(RootI13nComponent, {
+        contextTypes: ComponentSpecs.staticSpecs.contextTypes,
+        childContextTypes: ComponentSpecs.staticSpecs.childContextTypes
     });
+    
+    var componentName = Component.displayName || Component.name;
+    RootI13nComponent.displayName = options.displayName || ('RootI13n' + componentName);
+    hoistNonReactStatics(RootI13nComponent, Component);
 
     return RootI13nComponent;
 };
