+++ tests/functional/bootstrap.js
@@ -14,6 +14,6 @@ window.I13nButton = require('../../dist/components/I13nButton');
 window.I13nDiv = require('../../dist/components/I13nDiv');
 
 window.createI13nNode = require('../../dist/utils/createI13nNode');
-window.clickHandler = require('../../dist/utils/clickHandler');
+window.clickHandler = require('../../dist/libs/clickHandler');
 window.setupI13n = require('../../dist/utils/setupI13n');
 window.I13nUtils = require('../../dist/mixins/I13nUtils');
